Greatly speed up duospace-width
duospace-width() is getting really heavy use in Terminal::Widgets, to the point
of being a performance bottleneck on high-cardinality non-cacheable inputs.

This commit replaces the core of the implementation with a *much* faster NQP
version, making the overall function ~7x faster on large inputs.

If the SGR color escape stripping functionality is NOT needed, the core is
also exported directly as duospace-width-core(), which is ~10x faster than the
original implementation.